+++ pkgs/ffigen/lib/src/code_generator/library.dart
@@ -11,6 +11,9 @@ import 'package:yaml_edit/yaml_edit.dart';
 import '../code_generator.dart';
 import '../config_provider/config_types.dart';
 import '../visitor/ast.dart';
+import '../visitor/copy_methods_from_super_type.dart';
+import '../visitor/fill_method_dependencies.dart';
+import '../visitor/fix_overridden_methods.dart';
 import '../visitor/list_bindings.dart';
 
 import 'utils.dart';
@@ -41,7 +44,11 @@ class Library {
     List<String> nativeEntryPoints = const <String>[],
     this.objCBuiltInFunctions,
   }) {
-    _findBindings(bindings, sort);
+    _findBindings(bindings, sort, [
+      CopyMethodsFromSuperTypesVisitation.new,
+      FixOverriddenMethodsVisitation.new,
+      FillMethodDependenciesVisitation.new,
+    ]);
 
     final codeGenBindings =
         this.bindings.where((b) => b.generateBindings).toList();
@@ -106,12 +113,17 @@ class Library {
     );
   }
 
-  void _findBindings(List<Binding> original, bool sort) {
-    final visitation = ListBindingsVisitation();
-    Visitor(visitation).visitAll(original);
+  void _findBindings(List<Binding> roots, bool sort,
+      List<Visitation Function()> visitationBuidlers) {
+    for (final builder in visitationBuidlers) {
+      Visitor(builder()).visitAll(roots);
+    }
 
-    /// Save bindings.
+    final visitation = ListBindingsVisitation();
+    Visitor(visitation).visitAll(roots);
     bindings = visitation.bindings;
+
+    /// Sort bindings.
     if (sort) {
       bindings.sortBy((b) => b.name);
       for (final b in bindings) {
