+++ pkgs/ffigen/lib/src/visitor/fix_overridden_methods.dart
@@ -1,0 +1,44 @@
+// Copyright (c) 2024, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../code_generator.dart';
+
+import 'ast.dart';
+
+class FixOverriddenMethodsVisitation extends Visitation {
+  @override
+  void visitObjCInterface(ObjCInterface node) {
+    node.visitChildren(visitor);
+
+    // ObjC ignores nullability when deciding if an override for an inherited
+    // method is valid. But in Dart it's invalid to override a method and change
+    // it's return type from non-null to nullable, or its arg type from nullable
+    // to non-null. So in these cases we have to make the non-null type
+    // nullable, to avoid Dart compile errors.
+    for (var t = node.superType; t != null; t = t.superType) {
+      for (final method in node.methods) {
+        final superMethod = t.getMethod(method.originalName);
+        if (superMethod != null &&
+            !superMethod.isClassMethod &&
+            !method.isClassMethod) {
+          if (superMethod.returnType.typealiasType is! ObjCNullable &&
+              method.returnType.typealiasType is ObjCNullable) {
+            superMethod.returnType = ObjCNullable(superMethod.returnType);
+          }
+          final numArgs = method.params.length < superMethod.params.length
+              ? method.params.length
+              : superMethod.params.length;
+          for (var i = 0; i < numArgs; ++i) {
+            final param = method.params[i];
+            final superParam = superMethod.params[i];
+            if (superParam.type.typealiasType is ObjCNullable &&
+                param.type.typealiasType is! ObjCNullable) {
+              param.type = ObjCNullable(param.type);
+            }
+          }
+        }
+      }
+    }
+  }
+}
